# 03-Azure/01-04-AI/01_Agentic_SDLC/src/api-ts/api-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Express API with Swagger",
        "version": "1.0.0",
        "description": "REST API documentation using Swagger/OpenAPI"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/headquarters": {
            "get": {
                "summary": "Returns all headquarters",
                "tags": [
                    "Headquarters"
                ],
                "responses": {
                    "200": {
                        "description": "List of all headquarters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Headquarters"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new headquarters",
                "tags": [
                    "Headquarters"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Headquarters"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Headquarters created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Headquarters"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/headquarters/{id}": {
            "get": {
                "summary": "Get a headquarters by ID",
                "tags": [
                    "Headquarters"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Headquarters ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Headquarters found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Headquarters"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Headquarters not found"
                    }
                }
            },
            "put": {
                "summary": "Update a headquarters",
                "tags": [
                    "Headquarters"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Headquarters ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Headquarters"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Headquarters updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Headquarters"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Headquarters not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a headquarters",
                "tags": [
                    "Headquarters"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Headquarters ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Headquarters deleted successfully"
                    },
                    "404": {
                        "description": "Headquarters not found"
                    }
                }
            }
        },
...
        "/api/headquarters/{id}/label": {
            "get": {
                "summary": "Get headquarters label by ID",
                "tags": [
                    "Headquarters"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Headquarters ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Headquarters label created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "label": {
                                            "type": "string",
                                            "description": "Formatted location label"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Headquarters not found"
                    }
                }
            }
        },
...
        "/api/orders/{id}": {
            "get": {
                "summary": "Get an order by ID",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "put": {
                "summary": "Update an order",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete an order",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Order deleted successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/api/order-details": {
            "get": {
                "summary": "Returns all order details",
                "tags": [
                    "Order Details"
                ],
                "responses": {
                    "200": {
                        "description": "List of all order details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderDetail"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new order detail",
                "tags": [
                    "Order Details"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderDetail"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order detail created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/order-details/{id}": {
            "get": {
                "summary": "Get an order detail by ID",
                "tags": [
                    "Order Details"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order detail ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order detail found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order detail not found"
                    }
                }
            },
            "put": {
                "summary": "Update an order detail",
                "tags": [
                    "Order Details"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order detail ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/OrderDetail"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order detail updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order detail not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete an order detail",
                "tags": [
                    "Order Details"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order detail ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Order detail deleted successfully"
                    },
                    "404": {
                        "description": "Order detail not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Headquarters": {
                "type": "object",
                "required": [
                    "headquartersId",
                    "name"
                ],
                "properties": {
                    "headquartersId": {
                        "type": "integer",
                        "description": "The unique identifier for the headquarters"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the headquarters"
                    },
                    "address": {
                        "type": "string",
                        "description": "Main office address of the headquarters"
                    },
                    "phone": {
                        "type": "string",
                        "description": "Contact phone number for the headquarters"
                    },
                    "description": {
                        "type": "string",
                        "description": "Additional details about the headquarters"
                    },
                    "contactPerson": {
                        "type": "string",
                        "description": "Name of the primary contact person"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Contact email for the headquarters"
                    },
                    "city": {
                        "type": "string",
                        "description": "City where the headquarters is located"
                    },
                    "country": {
                        "type": "string",
                        "description": "Country where the headquarters is located"
                    },
                    "floorCount": {
                        "type": "integer",
                        "description": "Number of floors in the headquarters building"
                    },
                    "capacity": {
                        "type": "integer",
                        "description": "Total capacity of the headquarters"
                    }
                }
            },
            "Order": {
                "type": "object",
                "required": [
                    "orderId",
                    "branchId",
                    "orderDate"
                ],
                "properties": {
                    "orderId": {
                        "type": "integer",
                        "description": "The unique identifier for the order"
                    },
                    "branchId": {
                        "type": "integer",
                        "description": "The ID of the branch that placed the order"
                    },
                    "orderDate": {
                        "type": "string",
                        "format": "date-time",
                        "description": "The date and time when the order was placed"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the order"
                    },
                    "description": {
                        "type": "string",
                        "description": "Additional details about the order"
                    },
                    "status": {
                        "type": "string",
                        "description": "The current status of the order",
                        "enum": [
                            "pending",
                            "processing",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ]
                    }
                }
            },
            "OrderDetail": {
                "type": "object",
                "required": [
                    "orderDetailId",
                    "orderId",
                    "productId",
                    "quantity",
                    "unitPrice"
                ],
                "properties": {
                    "orderDetailId": {
                        "type": "integer",
                        "description": "The unique identifier for the order detail"
                    },
                    "orderId": {
                        "type": "integer",
                        "description": "The ID of the parent order"
                    },
                    "productId": {
                        "type": "integer",
                        "description": "The ID of the product ordered"
                    },
                    "quantity": {
                        "type": "integer",
                        "description": "The quantity of products ordered"
                    },
                    "unitPrice": {
                        "type": "number",
                        "format": "float",
                        "description": "The price per unit"
                    },
                    "notes": {
                        "type": "string",
                        "description": "Additional notes for the order detail"
                    }
                }
            },
...
        }
    },
...
}